WEBAPP-126: Wrap I18nProviderConfiguration in own React component
